Adding code to address issue 725, no waiting for core metadata
