+++ cmd/support-scheduler/main.go
@@ -4,7 +4,6 @@ import (
 	"flag"
 	"fmt"
 
-	"github.com/edgexfoundry/edgex-go/internal/pkg/startup"
 	"net/http"
 	"os"
 	"os/signal"
@@ -14,6 +13,7 @@ import (
 
 	"github.com/edgexfoundry/edgex-go"
 	"github.com/edgexfoundry/edgex-go/internal"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/startup"
 	"github.com/edgexfoundry/edgex-go/internal/pkg/usage"
 	"github.com/edgexfoundry/edgex-go/internal/support/scheduler"
 	"github.com/edgexfoundry/edgex-go/pkg/clients/logging"
@@ -36,19 +36,36 @@ func main() {
 	params := startup.BootParams{UseConsul: useConsul, UseProfile: useProfile, BootTimeout: internal.BootTimeoutDefault}
 	startup.Bootstrap(params, scheduler.Retry, logBeforeInit)
 
+	// ensure metadata is up
+	deps := make(chan string, 2)
+	go func(ch chan string) {
+		for {
+			select {
+			case m, ok := <-ch:
+				if ok {
+					scheduler.LoggingClient.Info(m)
+				} else {
+					return
+				}
+			}
+		}
+	}(deps)
+	scheduler.CheckStatus(params, scheduler.RetryService, deps)
+
+	// Load configuration data
 	ok := scheduler.Init(useConsul)
 	if !ok {
 		logBeforeInit(fmt.Errorf("%s: Service bootstrap failed!", internal.SupportSchedulerServiceKey))
 		return
 	}
 
-	scheduler.LoggingClient.Info("Service dependencies resolved...")
+	scheduler.LoggingClient.Info(fmt.Sprintf("Service dependencies resolved...%s", internal.SupportSchedulerServiceKey))
 	scheduler.LoggingClient.Info(fmt.Sprintf("Starting %s %s ", internal.SupportSchedulerServiceKey, edgex.Version))
 
 	// Bootstrap schedulers
 	err := scheduler.AddSchedulers()
-	if err != nil{
-		scheduler.LoggingClient.Error(fmt.Sprintf("Failed to load schedules and events %s",err.Error()))
+	if err != nil {
+		scheduler.LoggingClient.Error(fmt.Sprintf("Failed to load schedules and events %s", err.Error()))
 	}
 
 	http.TimeoutHandler(nil, time.Millisecond*time.Duration(scheduler.Configuration.Service.Timeout), "Request timed out")
@@ -68,7 +85,6 @@ func main() {
 	scheduler.Destruct()
 	scheduler.LoggingClient.Warn(fmt.Sprintf("terminating: %v", c))
 }
-
 
 func logBeforeInit(err error) {
 	scheduler.LoggingClient = logger.NewClient(internal.CoreCommandServiceKey, false, "", logger.InfoLog)
